Place the edition of parameters in a dedicated composable
